refactor: use active tense confirm instead of confirmation
