+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/brp/confirm/ConfirmBrpScreen.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.brp.confirmation
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.brp.confirm
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
