+++ features/select-doc/internal/src/test/snapshots/images/features.selectdoc.internal.brp.confirm_ConfirmBrpScreenKt_ConfirmBrpPreview_Light.png
