+++ features/select-doc/internal/src/test/snapshots/images/features.selectdoc.internal.passport.confirm_ConfirmDocumentScreenKt_PreviewConfirmDocumentScreen_Dark_NIGHT.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2da2cf68edd9b70b437f81c2e72ab4a1db7dfc830e491f8beb36861af47bb21a
+size 3118
