+++ features/select-doc/internal/src/test/snapshots/images/features.selectdoc.internal.passport.confirm_ConfirmDocumentScreenKt_PreviewConfirmDocumentScreen_Light.png
